Rewrite setup, add command
