+++ content-copy-tool/contentcopytool/lib/bookmap.py
