+++ content-copy-tool/contentcopytool/lib/http_util.py
