+++ content-copy-tool/contentcopytool/lib/operation_objects.py
