+++ content-copy-tool/contentcopytool/lib/util.py
