--- content-copy-tool/setup.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-easy_install --user requests
-easy_install --user requests[security]
-
-chmod +x content-copy.py
